backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
pydantic-settings==2.9.1
pydantic[email]==2.11.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pinecone==6.0.2
ollama==0.4.8
google-generativeai==0.8.5
PyMuPDF==1.25.5
python-dotenv==1.1.0
